If applied, this tidies up JS Docs, adds new UI test, organises & groups files
