+++ views/case_details/split_case/about-new-split-case.njk
@@ -32,7 +32,7 @@
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
 
-      <form method="POST" action="/cases/{{ caseReference }}/about-new-case">
+      <form method="POST" action="/cases/{{ caseReference }}/about-new-split-case">
 
         <!-- CSRF Protection -->
         <input type="hidden" name="_csrf" value="{{ csrfToken }}"/>
