+++ views/case_details/split_case/split-this-case.njk
